pyproject.toml
[project]
name = "gtr-task2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.119.1",
    "firecrawl-py>=4.5.0",
    "langchain-core>=1.0.0",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-postgres>=3.0.0",
    "langgraph-cli[inmem]>=0.4.4",
    "langsmith>=0.4.37",
    "psycopg[binary]>=3.2.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "rich>=14.2.0",
    "sqlmodel>=0.0.27",
]
